requirements.txt
discord.py

cryptography

requests

Pillow

psutil

selenium